// src/apiclient/__generated__/ScanCfnTemplate.ts
export interface ScanCfnTemplate_scanCfnTemplateExt_scanMeta {
  __typename: "ScanMeta";
  timestamp: string;
  scanId: string;
  portalUrl: string;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_sideEffectsResult {
  __typename: "SideEffectsResult";
  success: boolean;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations_logicalResource {
  __typename: "LogicalResource";
  line: number;
  name: string;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations_policyStatement_capability {
  __typename: "Capability";
  title: string;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations_policyStatement {
  __typename: "MustImplementCapabilityPolicyStatement";
  capability: ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations_policyStatement_capability;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations {
  __typename: "ComplianceObservation";
  logicalResource: ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations_logicalResource;
  policyStatement: ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations_policyStatement;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_policyStatement_capability {
  __typename: "Capability";
  title: string;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_policyStatement {
  __typename: "MustImplementCapabilityPolicyStatement";
  capability: ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_policyStatement_capability;
}
// ...
export interface ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations {
  __typename: "ViolationObservation";
  policyStatement: ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_policyStatement;
  trivialRemediation: ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_trivialRemediation | null;
  nonTrivialRemediation: ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_nonTrivialRemediation[];
  logicalResource: ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations_logicalResource;
}

export interface ScanCfnTemplate_scanCfnTemplateExt_results {
  __typename: "CfnScanResult";
  filePath: string;
  error: string | null;
  complianceObservations: ScanCfnTemplate_scanCfnTemplateExt_results_complianceObservations[];
  violationObservations: ScanCfnTemplate_scanCfnTemplateExt_results_violationObservations[];
}

export interface ScanCfnTemplate_scanCfnTemplateExt {
  __typename: "ScanCfnResultType";
  scanMeta: ScanCfnTemplate_scanCfnTemplateExt_scanMeta;
  sideEffectsResult: ScanCfnTemplate_scanCfnTemplateExt_sideEffectsResult | null;
  results: ScanCfnTemplate_scanCfnTemplateExt_results[];
}